+++ crates/dialogs/src/lib.rs
@@ -5,4 +5,5 @@ mod handle;
 mod presets;
 
 pub use error::{Error, Result};
+pub use global::init;
 pub use handle::{Dialog, DialogMsg};
