+++ crates/rpc/Cargo.toml
@@ -15,7 +15,6 @@ iron-db = { workspace = true }
 iron-dialogs = { workspace = true }
 iron-networks = { workspace = true }
 iron-wallets = { workspace = true }
-iron-sync-anvil = { workspace = true }
 
 ethers = { workspace = true }
 tokio = { workspace = true }
