+++ crates/settings/Cargo.toml
@@ -12,7 +12,6 @@ authors.workspace = true
 iron-types = { workspace = true }
 iron-db = { workspace = true }
 ethers = { workspace = true }
-iron-sync-anvil = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tauri = { workspace = true }
