+++ crates/wallets/Cargo.toml
@@ -11,7 +11,6 @@ authors.workspace = true
 
 [dependencies]
 iron-types = { workspace = true }
-iron-globals = { workspace = true }
 iron-dialogs = { workspace = true }
 iron-crypto = { workspace = true }
 iron-broadcast = { workspace = true }
